refactor(frontend): update folder responsabilities
